experiments/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
accelerate=0.21.0=pyhd8ed1ab_0
aiohttp=3.8.3=py311h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
arrow-cpp=11.0.0=h374c478_1
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=py311h06a4308_0
attrs=22.1.0=py311h06a4308_0
aws-c-common=0.6.8=h5eee18b_1
aws-c-event-stream=0.1.6=h6a678d5_6
aws-checksums=0.1.11=h5eee18b_2
aws-sdk-cpp=1.8.185=h721c034_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
black=23.7.0=py311h38be061_1
blas=1.0=mkl
boost-cpp=1.73.0=h7f8727e_12
bottleneck=1.3.5=py311hbed6279_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.11.17=py311h06a4308_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py311h06a4308_0
colorama=0.4.6=py311h06a4308_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.5=py311hdb19cb5_0
cryptography=41.0.2=py311h22a60cf_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
curl=8.2.1=hdbd6064_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
datasets=2.12.0=py311h06a4308_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.8=py311hb755f60_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.6=py311h06a4308_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
einops=0.6.1=pyhd8ed1ab_0
et_xmlfile=1.1.0=py311h06a4308_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=hcb278e6_1
filelock=3.9.0=py311h06a4308_0
fire=0.5.0=pyhd8ed1ab_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py311h5eee18b_0
fsspec=2023.12.0=pypi_0
gdbm=1.18=hd4cb3f1_4
gettext=0.21.0=hf68c758_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
git=2.40.1=pl5340h36fbf9e_1
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py311h06a4308_0
glib=2.69.1=he621ea3_2
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
grpc-cpp=1.48.2=he1ff14a_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
huggingface-hub=0.19.4=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py311h06a4308_0
importlib-metadata=6.0.0=py311h06a4308_0
importlib_metadata=6.0.0=hd3eb1b0_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.25.1=pyh71e2992_0
ipython=8.14.0=pyh41d4057_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py311h06a4308_0
joblib=1.2.0=py311h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h38be061_0
kiwisolver=1.4.4=py311h6a678d5_0
krb5=1.20.1=h143b758_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.73.0=h28710b8_12
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.1.12=0
libcurand=10.3.3.129=0
libcurl=8.2.1=h251f7ec_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=13.1.0=he5830b7_0
libllvm10=10.0.1=hbcb73fb_5
libnghttp2=1.52.0=h2d74bed_1
libnpp=11.8.0.86=0
libnsl=2.0.0=h5eee18b_0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.42.0=h2797004_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.1.0=hfd8a6a1_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.0=h6a678d5_2
libuuid=2.38.1=h0b41bf4_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzlib=1.2.13=hd590300_5
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py311h5eee18b_0
matplotlib=3.7.2=py311h38be061_0
matplotlib-base=3.7.2=py311h54ef318_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.6=py311ha02d727_1
mkl_random=1.2.2=py311ha02d727_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.0.2=py311h5eee18b_0
multiprocess=0.70.14=py311h06a4308_0
munkres=1.1.4=py_0
mypy_extensions=0.4.3=py311h06a4308_1
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=py311h06a4308_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.4=py311h65dcdc2_1
numpy=1.25.0=py311h08b1b3b_0
numpy-base=1.25.0=py311hf175353_0
openai=0.27.8=pyhd8ed1ab_0
openpyxl=3.0.10=py311h5eee18b_0
openssl=3.2.0=hd590300_1
orc=1.7.4=hb3bc3d3_1
packaging=23.0=py311h06a4308_0
pandas=1.5.3=py311hba01205_0
pandas-stubs=1.5.3.230203=py311h06a4308_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.10.3=py311h06a4308_0
pathtools=0.1.2=pyhd3eb1b0_1
patsy=0.5.3=py311h06a4308_0
pcre=8.45=h295c915_0
pcre2=10.42=hebb0a14_0
perl=5.34.0=h5eee18b_2
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311h6a678d5_0
pip=23.3.1=py311h06a4308_0
platformdirs=2.5.2=py311h06a4308_0
plotly=5.15.0=pyhd8ed1ab_0
ply=3.11=py311h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=3.20.3=py311h6a678d5_0
psutil=5.9.0=py311h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py311hd8e8d9b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=23.2.0=py311h06a4308_0
pyparsing=3.0.9=py311h06a4308_0
pyqt=5.15.7=py311h6a678d5_0
pyqt5-sip=12.11.0=py311h6a678d5_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.4=hab00c5b_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-xxhash=2.0.2=py311h5eee18b_1
python_abi=3.11=3_cp311
pytorch=2.0.1=py3.11_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2022.7=py311h06a4308_0
pyyaml=6.0=py311h5eee18b_1
pyzmq=25.1.1=py311h75c88c4_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py311h5eee18b_0
requests=2.31.0=py311h06a4308_0
responses=0.13.3=pyhd3eb1b0_0
sacremoses=0.0.43=pyhd3eb1b0_0
safetensors=0.3.1=py311h46250e7_0
scikit-learn=1.2.2=py311h6a678d5_1
scipy=1.10.1=py311h08b1b3b_1
seaborn=0.12.2=hd8ed1ab_0
seaborn-base=0.12.2=pyhd8ed1ab_0
sentencepiece=0.1.99=py311hdb19cb5_0
sentry-sdk=1.9.0=py311h06a4308_0
setproctitle=1.2.2=py311h5eee18b_0
setuptools=68.0.0=py311h06a4308_0
sip=6.6.2=py311h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=py311hf4808d0_0
sympy=1.11.1=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.2=py311h06a4308_0
termcolor=2.3.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tiktoken=0.4.0=py311h1069841_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.0=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torchtriton=2.0.0=py311
tornado=6.3.2=py311h5eee18b_0
tqdm=4.65.0=py311h92b7b1e_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.35.2=pypi_0
types-pytz=2022.4.0.0=py311h06a4308_1
typing-extensions=4.7.1=py311h06a4308_0
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wandb=0.15.8=pyhd8ed1ab_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py311h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py311h5eee18b_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.11.0=py311h06a4308_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hc292b87_0
